blitz_python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "blitz-python"                    # distribution name (pip install blitz-python)
version = "0.1.0"
description = "Serial communication library for MCU ↔ PC"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.7"
authors = [
  { name = "Krishnapranav B" }
]
license = { text = "MIT" }
dependencies = [
  "pyserial"
]

[project.urls]
Homepage = "https://github.com/yourusername/blitz"
Repository = "https://github.com/yourusername/blitz"
